fix bug
update user sign in, sign out, reset password and resend email features

+++ Backend/GroupManagement/src/main/java/com/vti/repository/RegistrationUserTokenRepository.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.vti.entity.RegistrationUserToken;
 
 
-public interface RegistrationUserTokenRepository extends JpaRepository<RegistrationUserToken, Integer> {
+public interface RegistrationUserTokenRepository extends JpaRepository<RegistrationUserToken, Short> {
 
 	public RegistrationUserToken findByToken(String token);
